+++ R/np.kernel.R
@@ -454,7 +454,7 @@ npksum.default <-
 
     if (internal.power12) {
       myout <-
-        .Call("C_np_kernelsum_power12",
+        .np_with_nn_radius_context(.Call("C_np_kernelsum_power12",
               asDouble(tuno), asDouble(tord), asDouble(tcon),
               asDouble(tydat), asDouble(weights),
               asDouble(euno), asDouble(eord), asDouble(econ),
@@ -467,10 +467,10 @@ npksum.default <-
               as.integer(nkw),
               as.double(cker.bounds.c$lb),
               as.double(cker.bounds.c$ub),
-              PACKAGE="np")[return.names]
+              PACKAGE="np"), continuous.names = bws[["xnames", exact = TRUE]][bws[["icon", exact = TRUE]]])[return.names]
     } else {
       myout <-
-        .Call("C_np_kernelsum",
+        .np_with_nn_radius_context(.Call("C_np_kernelsum",
               asDouble(tuno), asDouble(tord), asDouble(tcon),
               asDouble(tydat), asDouble(weights),
               asDouble(euno), asDouble(eord), asDouble(econ),
@@ -483,7 +483,7 @@ npksum.default <-
               as.integer(nkw),
               as.double(cker.bounds.c$lb),
               as.double(cker.bounds.c$ub),
-              PACKAGE="np")[return.names]
+              PACKAGE="np"), continuous.names = bws[["xnames", exact = TRUE]][bws[["icon", exact = TRUE]]])[return.names]
     }
 
     if (dim.out[1] > 1){