+++ R/np.reghat.R
@@ -1081,7 +1081,7 @@ npreghat <-
                    sort(c("mode", "coordinate", "response.ready", "pivotal",
                           "null.mean", "residual.pool"))))
       stop("invalid private npsigtest tile payload", call. = FALSE)
-    return(.Call(
+    return(.np_with_nn_radius_context(.Call(
       "C_np_regression_lp_sigtest_conditional_ctx",
       tuno,
       tord,
@@ -1115,10 +1115,10 @@ npreghat <-
       as.double(sigtest$null.mean),
       as.double(sigtest$residual.pool),
       PACKAGE = "np"
-    ))
+    ), continuous.names = bws[["xnames", exact = TRUE]][bws[["icon", exact = TRUE]]]))
   }
 
-  .Call(
+  .np_with_nn_radius_context(.Call(
     "C_np_regression_lp_apply_conditional_ctx",
     tuno,
     tord,
@@ -1146,7 +1146,7 @@ npreghat <-
     as.logical(no.ex),
     as.logical(leave.one.out),
     PACKAGE = "np"
-  )
+  ), continuous.names = bws[["xnames", exact = TRUE]][bws[["icon", exact = TRUE]]])
 }
 
 .npreghat_exact_lp_matrix_from_regression_core <- function(bws,
@@ -1564,7 +1564,7 @@ npreghat <-
     integer(1L)
   }
 
-  myout <- .Call(
+  myout <- .np_with_nn_radius_context(.Call(
     "C_np_regression",
     asDouble(tuno), asDouble(tord), asDouble(tcon), as.double(tydat),
     asDouble(euno), asDouble(eord), asDouble(econ), as.double(double()),
@@ -1585,7 +1585,7 @@ npreghat <-
     as.double(cker.bounds.c$lb),
     as.double(cker.bounds.c$ub),
     PACKAGE = "np"
-  )
+  ), continuous.names = bws[["xnames", exact = TRUE]][bws[["icon", exact = TRUE]]])
 
   out <- list(mean = as.double(myout$mean))